Add files via upload
it has a winner and loser function but it will not show if there is a draw so kindly use the replay button. Another concern is when you press a zone like Zone 1 it will make an X but when you press it again it will make the text O, the concern is that the text that is contained by the label is changing and it is not locked when you press the zone twice or thrice.